pyproject.toml
[project]
name = "courts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.10.5",
    "bs4>=0.0.2",
    "celery>=5.4.0",
    "docx2txt>=0.8",
    "flower>=2.0.1",
    "opencv-python>=4.10.0.84",
    "pandas>=2.2.2",
    "pdfplumber>=0.11.4",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "striprtf>=0.0.26",
    "tensorflow>=2.17.0",
    "tf-keras>=2.17.0",
]

[tool.uv]
dev-dependencies = [
    "pytest-asyncio>=0.24.0",
    "pytest-vcr>=1.0.2",
    "pytest>=8.3.3",
    "ruff>=0.6.5",
    "vcrpy>=6.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
